Move large scnery entry
