+++ src/openrct2-ui/windows/Scenery.cpp
@@ -22,6 +22,7 @@
 #include <openrct2/network/network.h>
 #include <openrct2/object/BannerObject.h>
 #include <openrct2/object/FootpathItemObject.h>
+#include <openrct2/object/LargeSceneryObject.h>
 #include <openrct2/object/ObjectList.h>
 #include <openrct2/object/SmallSceneryObject.h>
 #include <openrct2/object/WallObject.h>
