+++ src/openrct2-ui/windows/Sign.cpp
@@ -19,6 +19,7 @@
 #include <openrct2/config/Config.h>
 #include <openrct2/localisation/Localisation.h>
 #include <openrct2/localisation/StringIds.h>
+#include <openrct2/object/LargeSceneryObject.h>
 #include <openrct2/object/WallObject.h>
 #include <openrct2/sprites.h>
 #include <openrct2/world/Banner.h>
