+++ src/openrct2/actions/LargeSceneryRemoveAction.cpp
@@ -17,6 +17,7 @@
 #include "../localisation/Localisation.h"
 #include "../localisation/StringIds.h"
 #include "../management/Finance.h"
+#include "../object/LargeSceneryObject.h"
 #include "../ride/Ride.h"
 #include "../world/Park.h"
 #include "../world/Scenery.h"
