+++ src/openrct2/actions/LargeScenerySetColourAction.cpp
@@ -11,6 +11,7 @@
 
 #include "../OpenRCT2.h"
 #include "../management/Finance.h"
+#include "../object/LargeSceneryObject.h"
 #include "../world/Scenery.h"
 
 LargeScenerySetColourAction::LargeScenerySetColourAction(
