+++ src/openrct2/entity/Guest.cpp
@@ -29,6 +29,7 @@
 #include "../management/NewsItem.h"
 #include "../network/network.h"
 #include "../object/FootpathItemObject.h"
+#include "../object/LargeSceneryObject.h"
 #include "../object/MusicObject.h"
 #include "../object/WallObject.h"
 #include "../peep/GuestPathfinding.h"
