+++ src/openrct2/interface/Viewport.cpp
@@ -22,6 +22,7 @@
 #include "../entity/Guest.h"
 #include "../entity/PatrolArea.h"
 #include "../entity/Staff.h"
+#include "../object/LargeSceneryObject.h"
 #include "../object/SmallSceneryObject.h"
 #include "../object/WallObject.h"
 #include "../paint/Paint.h"
