+++ src/openrct2/world/MapAnimation.cpp
@@ -14,6 +14,7 @@
 #include "../entity/EntityList.h"
 #include "../entity/Peep.h"
 #include "../interface/Viewport.h"
+#include "../object/LargeSceneryObject.h"
 #include "../object/SmallSceneryObject.h"
 #include "../object/StationObject.h"
 #include "../object/WallObject.h"
