+++ src/openrct2/world/TileInspector.cpp
@@ -17,6 +17,7 @@
 #include "../interface/Window.h"
 #include "../interface/Window_internal.h"
 #include "../localisation/Localisation.h"
+#include "../object/LargeSceneryObject.h"
 #include "../ride/Station.h"
 #include "../ride/Track.h"
 #include "../ride/TrackData.h"
